DNAConsole tgui interface refactor (#59407)
* Improve tgui routing to allow interface subdirectories

* Reverse order of interface file resolution, to be consistent with common expectation (top-level .tsx, top-level .js, directory .tsx, directory.js) and the way Webpack checks

* Refactor DnaConsole tgui interface into separate files
